test examen

+++ src/main/java/tn/esprit/spring/entities/Employe.java
@@ -26,7 +26,7 @@ public class Employe implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	private String prenom;
 	
@@ -70,11 +70,13 @@ public class Employe implements Serializable {
 		this.role = role;
 	}
 	
-	public int getId() {
+	
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
