+++ src/main/java/tn/esprit/spring/repository/TimesheetRepository.java
@@ -16,7 +16,7 @@ import tn.esprit.spring.entities.TimesheetPK;
 public interface TimesheetRepository extends CrudRepository<Timesheet, Integer> {
 
 	@Query("select DISTINCT m from Mission m join m.timesheets t join t.employe e where e.id=:employeId")
-	public List<Mission> findAllMissionByEmployeJPQL(@Param("employeId")int employeId);
+	public List<Mission> findAllMissionByEmployeJPQL(@Param("employeId")long employeId);
 	
 	@Query("select DISTINCT e from Employe e "
 				+ "join e.timesheets t "
